[cocoemu-server] Change Handler and CdM-16 so we can use different adapters
